adds different scopes for datadirs
In order to accomplish this I generalized the creation of datadirs,
and then provided aliases for the old fixtures of the same name.

This adds a number of fixtures to the namespace which are the same as
before but prefixed by the pytest scope they apply to,
e.g. `function_shared_datadir`.  Which is the same as the old
`shared_datadir`.

This also opens up for more complex (i.e. non-hardcoded) usage
scenarios, although I have just stuck to the same behaviors as before.

The only difference perhaps is that previously there was a behavior
where if the original data dir was not a directory at all a directory
in the temp dir was created with that name, although I didn't quite
understand why that was needed or what pattern it supports.