reimplements factory and fixtures
I was confused on how the factory should work before. And found a
problem with the use of a closure using the nonlocal. When I tested it
before I was testing only a single module which got the
'original_datadir' right, but when I ran the full suite with others,
the first closure value for 'original_datadir' would get carried to
the rest of them and be wrong.

In this approach the factory is a POPO and works similar to the
`tmp_path_factory` class.

Generated the same fixtures as before.